+++ src/test/java/pages/HomePage.java
@@ -1,0 +1,43 @@
+package pages;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.Keys;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.Wait;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.time.Duration;
+
+import static org.openqa.selenium.By.cssSelector;
+
+public class HomePage extends BasePage {
+
+    public HomePage(WebDriver givenDriver) {
+        super(givenDriver);
+    }
+
+    public String getUpdatedPlaylistMsg() {
+        WebElement playListMsg = driver.findElement(cssSelector("div.success.show"));
+        return playListMsg.getText();
+    }
+
+    public void selectPlayList(String playlist) {
+        Wait<WebDriver> wait = new WebDriverWait(driver, Duration.ofSeconds(3));
+        WebElement selectPlayListElement = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector(".playlist:nth-child(3)")));
+
+        Actions action = new Actions(driver);
+
+        action.
+                doubleClick(selectPlayListElement).
+                doubleClick(selectPlayListElement).
+                sendKeys(selectPlayListElement, Keys.DELETE).perform();
+
+        action.
+                sendKeys(selectPlayListElement, playlist).
+                sendKeys(Keys.ENTER).perform();
+
+    }
+}
