+++ src/test/resources/TestNG.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="homework22">
+<suite name="homework23">
     <parameter name="baseUrl" value="https://qa.koel.app/"/>
-    <test name="homework22" preserve-order="false">
+    <test name="homework23" preserve-order="false">
         <classes>
             <class name="BaseTest2"/>
-            <class name="HomeWork22">
+            <class name="HomeWork23">
                 <methods>
                     <include name="renamePlaylist"/>
                 </methods>
